Compare downloaded files against the manifest
When downloading files in the loop, check that the set of downloaded
files matches what's expected in the manifest. All files were downloaded,
stop the loop. Also add a check that all files were successfully
downloaded in the main method, and exit with an error message if they
were not.