Combine both WDL tasks into one
It doesn't make sense to have two separate tasks here. Because we
need both to download the files and then to copy them to the final
destination on one VM, we should only run one task. Otherwise the
first task may have to have outputs, which would be stored in the
submissions directory (and thus be duplicates).